break custom search dialog into multiple modules
