+++ package.json
@@ -16,6 +16,8 @@
 		"validate-env": "node -e \"require('./lib/env').validateEnv()\""
 	},
 	"dependencies": {
+		"@radix-ui/react-dialog": "^1.1.14",
+		"@radix-ui/react-visually-hidden": "^1.2.3",
 		"cmdk": "^1.1.1",
 		"fumadocs-core": "15.2.6",
 		"fumadocs-mdx": "11.5.8",
